+++ src/Bridge/Doctrine/Orm/CollectionDataProvider.php
@@ -20,7 +20,6 @@ use ApiPlatform\Core\DataProvider\ContextAwareCollectionDataProviderInterface;
 use ApiPlatform\Core\DataProvider\RestrictedDataProviderInterface;
 use ApiPlatform\Core\Exception\RuntimeException;
 use Doctrine\Common\Persistence\ManagerRegistry;
-use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\ORM\EntityManagerInterface;
 
 /**
@@ -56,7 +55,7 @@ class CollectionDataProvider implements ContextAwareCollectionDataProviderInterf
      */
     public function getCollection(string $resourceClass, string $operationName = null, array $context = [])
     {
-        /** @var ObjectManager $manager */
+        /** @var EntityManagerInterface $manager */
         $manager = $this->managerRegistry->getManagerForClass($resourceClass);
 
         $repository = $manager->getRepository($resourceClass);